requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.7
httpx==0.25.2
python-multipart==0.0.6
markdown==3.5.1
bleach==6.1.0
python-dotenv==1.0.0